feat: Initial color management support
Managing NCurses colors is a bit nuanced. First of all, the short-based
Color pairs (init-pair) only support up to 255 pairs, even when a
terminal reports supporting more--this is due to attribute sizes that
ncurses respects for compatibility.

init-extended-pair can initialize pairs greater than 255 for terminals
that support them, but such colors can only be set using ncurses
extensions to the attr-set style function. Legacy functions like attron
only operate on the old pairs that are limited to 255.

So, for rich color support, you generally want to provide the extension
pointer argument to attr-set. I've accounted for some of these nuances
but not all of them--for instance, we need to check if the user has a
version of ncurses that supports the extended color routines.

Furthermore, our account of colors isn't truly robust until we have a
framework for dealing with attributes more generally.