feat: Add wrappers for char and str
These macros enable users to print characters (with or without
properties) to the screen, along with strings. Furthermore, I've added a
short example demonstrating their usage.

We only wrap the NCurses functions that take window arguments, not the
stdscr macros, as we pass the stdscr component to retrieve the stdscr
window (see a few commits back).